Refactor split bill functionality and integrate LIFF service for LINE login; update ESLint configuration and package dependencies
