+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@line/liff": "^2.26.1",
     "@supabase/supabase-js": "^2.49.4",
     "@types/three": "^0.176.0",
     "firebase": "^11.6.0",
@@ -20,7 +21,8 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hot-toast": "^2.5.2",
-    "three": "^0.176.0"
+    "three": "^0.176.0",
+    "zustand": "^5.0.5"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
